Adds compact view for product listings
Implements a compact view for product listings, primarily to be used in release lists.

This includes adjustments to thumbnail sizes and the removal of the description and date in the compact view, allowing for a more condensed presentation. It also adds support for displaying related releases within the product details when in fullscreen mode.